rename /namespaces/ -> /resources/ in the admin module
