+++ src/admin/resources/projects.ts
